figure handling
figures can now be run even if the models are not (formerly, if models were skipped they were stripped from the data file and could thus not be referred to in the figure step). Added an Enable field for figures. Fixed a bug where the figure code was outside the loop and thus only the last column would run.